+++ awscli/customizations/argrename.py
@@ -44,12 +44,30 @@ ARGUMENT_RENAMES = {
     'codepipeline.delete-custom-action-type.version': 'action-version',
 }
 
+# Same format as ARGUMENT_RENAMES, but instead of renaming the arguments,
+# an alias is created to the original arugment and marked as undocumented.
+# This is useful when you need to change the name of an argument but you
+# still need to support the old argument.
+HIDDEN_ALIASES = {
+    'cognito-identity.create-identity-pool.open-id-connect-provider-arns':
+        'open-id-connect-provider-ar-ns',
+    'storagegateway.describe-tapes.tape-arns': 'tape-ar-ns',
+    'storagegateway.describe-tape-archives.tape-arns': 'tape-ar-ns',
+    'storagegateway.describe-vtl-devices.vtl-device-arns': 'vtl-device-ar-ns',
+    'storagegateway.describe-cached-iscsi-volumes.volume-arns': 'volume-ar-ns',
+    'storagegateway.describe-stored-iscsi-volumes.volume-arns': 'volume-ar-ns',
+}
+
 
 def register_arg_renames(cli):
     for original, new_name in ARGUMENT_RENAMES.items():
         event_portion, original_arg_name = original.rsplit('.', 1)
         cli.register('building-argument-table.%s' % event_portion,
                      rename_arg(original_arg_name, new_name))
+    for original, new_name in HIDDEN_ALIASES.items():
+        event_portion, original_arg_name = original.rsplit('.', 1)
+        cli.register('building-argument-table.%s' % event_portion,
+                     hidden_alias(original_arg_name, new_name))
 
 
 def rename_arg(original_arg_name, new_name):
@@ -57,3 +75,10 @@ def rename_arg(original_arg_name, new_name):
         if original_arg_name in argument_table:
             utils.rename_argument(argument_table, original_arg_name, new_name)
     return _rename_arg
+
+
+def hidden_alias(original_arg_name, alias_name):
+    def _alias_arg(argument_table, **kwargs):
+        if original_arg_name in argument_table:
+            utils.make_hidden_alias(argument_table, original_arg_name, alias_name)
+    return _alias_arg