+++ tests/unit/customizations/s3/test_utils.py
@@ -24,6 +24,7 @@ from awscli.customizations.s3.utils import create_warning
 from awscli.customizations.s3.utils import human_readable_size
 from awscli.customizations.s3.utils import human_readable_to_bytes
 from awscli.customizations.s3.utils import MAX_SINGLE_UPLOAD_SIZE
+from awscli.customizations.s3.utils import set_file_utime, SetFileUtimeError
 
 
 def test_human_readable_size():
@@ -291,20 +292,26 @@ class TestBucketList(unittest.TestCase):
     def test_list_objects(self):
         now = mock.sentinel.now
         self.client.get_paginator.return_value.paginate = self.fake_paginate
+        individual_response_elements = [
+            {'LastModified': '2014-02-27T04:20:38.000Z',
+             'Key': 'a', 'Size': 1},
+            {'LastModified': '2014-02-27T04:20:38.000Z',
+                 'Key': 'b', 'Size': 2},
+            {'LastModified': '2014-02-27T04:20:38.000Z',
+                 'Key': 'c', 'Size': 3}
+        ]
         self.responses = [
-            {'Contents': [
-                {'LastModified': '2014-02-27T04:20:38.000Z',
-                 'Key': 'a', 'Size': 1},
-                {'LastModified': '2014-02-27T04:20:38.000Z',
-                 'Key': 'b', 'Size': 2}]},
-            {'Contents': [
-                {'LastModified': '2014-02-27T04:20:38.000Z',
-                 'Key': 'c', 'Size': 3}]}
+            {'Contents': individual_response_elements[0:2]},
+            {'Contents': [individual_response_elements[2]]}
         ]
         lister = BucketLister(self.client, self.date_parser)
         objects = list(lister.list_objects(bucket='foo'))
-        self.assertEqual(objects, [('foo/a', 1, now), ('foo/b', 2, now),
-                                   ('foo/c', 3, now)])
+        self.assertEqual(objects,
+            [('foo/a', individual_response_elements[0]),
+             ('foo/b', individual_response_elements[1]),
+             ('foo/c', individual_response_elements[2])])
+        for individual_response in individual_response_elements:
+            self.assertEqual(individual_response['LastModified'], now)
 
     def test_urlencoded_keys(self):
         # In order to workaround control chars being in key names,
@@ -313,28 +320,34 @@ class TestBucketList(unittest.TestCase):
         # in bar.txt:
         now = mock.sentinel.now
         self.client.get_paginator.return_value.paginate = self.fake_paginate
+        individual_response_element = {
+            'LastModified': '2014-02-27T04:20:38.000Z',
+            'Key': 'bar%0D.txt', 'Size': 1}
         self.responses = [
-            {'Contents': [
-                {'LastModified': '2014-02-27T04:20:38.000Z',
-                 'Key': 'bar%0D.txt', 'Size': 1}]},
+            {'Contents': [individual_response_element]}
         ]
         lister = BucketLister(self.client, self.date_parser)
         objects = list(lister.list_objects(bucket='foo'))
         # And note how it's been converted to '\r'.
-        self.assertEqual(objects, [('foo/bar\r.txt', 1, now)])
+        self.assertEqual(
+            objects, [('foo/bar\r.txt', individual_response_element)])
+        self.assertEqual(individual_response_element['LastModified'], now)
 
     def test_urlencoded_with_unicode_keys(self):
         now = mock.sentinel.now
         self.client.get_paginator.return_value.paginate = self.fake_paginate
+        individual_response_element = {
+            'LastModified': '2014-02-27T04:20:38.000Z',
+            'Key': '%E2%9C%93', 'Size': 1}
         self.responses = [
-            {'Contents': [
-                {'LastModified': '2014-02-27T04:20:38.000Z',
-                 'Key': '%E2%9C%93', 'Size': 1}]},
+            {'Contents': [individual_response_element]}
         ]
         lister = BucketLister(self.client, self.date_parser)
         objects = list(lister.list_objects(bucket='foo'))
         # And note how it's been converted to '\r'.
-        self.assertEqual(objects, [(u'foo/\u2713', 1, now)])
+        self.assertEqual(
+            objects, [(u'foo/\u2713', individual_response_element)])
+        self.assertEqual(individual_response_element['LastModified'], now)
 
 
 class TestScopedEventHandler(unittest.TestCase):
@@ -382,5 +395,28 @@ class TestGetFileStat(unittest.TestCase):
                     get_file_stat('myfilename.txt')
 
 
-if __name__ == "__main__":
-    unittest.main()
+class TestSetsFileUtime(unittest.TestCase):
+
+    def test_successfully_sets_utime(self):
+        now = datetime.datetime.now(tzlocal())
+        epoch_now = time.mktime(now.timetuple())
+        with temporary_file('w') as f:
+            set_file_utime(f.name, epoch_now)
+            _, update_time = get_file_stat(f.name)
+            self.assertEqual(time.mktime(update_time.timetuple()), epoch_now)
+
+    def test_throws_more_relevant_error_when_errno_1(self):
+        now = datetime.datetime.now(tzlocal())
+        epoch_now = time.mktime(now.timetuple())
+        with mock.patch('os.utime') as utime_mock:
+            utime_mock.side_effect = OSError(1, '')
+            with self.assertRaises(SetFileUtimeError):
+                set_file_utime('not_real_file', epoch_now)
+
+    def test_passes_through_other_os_errors(self):
+        now = datetime.datetime.now(tzlocal())
+        epoch_now = time.mktime(now.timetuple())
+        with mock.patch('os.utime') as utime_mock:
+            utime_mock.side_effect = OSError(2, '')
+            with self.assertRaises(OSError):
+                set_file_utime('not_real_file', epoch_now)